+++ Source/ReactPhysics/Private/engine/Island.cpp
@@ -30,7 +30,7 @@
 using namespace reactphysics3d;
 
 // Constructor
-Island::Island(uint nbMaxBodies, uint nbMaxContactManifolds, MemoryManager& memoryManager)
+Island::Island(uint8 nbMaxBodies, uint8 nbMaxContactManifolds, MemoryManager& memoryManager)
        : mBodies(nullptr), mContactManifolds(nullptr), mNbBodies(0), mNbContactManifolds(0) {
 
     // Allocate memory for the arrays on the single frame allocator
