+++ Source/ReactPhysics/Public/reactphysics3d/collision/ContactPair.h
@@ -48,7 +48,7 @@ struct ContactPair {
         uint64 pairId;
 
         /// Indices of the potential contact manifolds
-        List<uint> potentialContactManifoldsIndices;
+        List<uint8> potentialContactManifoldsIndices;
 
         /// Entity of the first body of the contact
         Entity body1Entity;
@@ -66,19 +66,19 @@ struct ContactPair {
         bool isAlreadyInIsland;
 
         /// Index of the contact pair in the array of pairs
-        uint contactPairIndex;
+        uint8 contactPairIndex;
 
         /// Index of the first contact manifold in the array
-        uint contactManifoldsIndex;
+        uint8 contactManifoldsIndex;
 
         /// Number of contact manifolds
         int8 nbContactManifolds;
 
         /// Index of the first contact point in the array of contact points
-        uint contactPointsIndex;
+        uint8 contactPointsIndex;
 
         /// Total number of contact points in all the manifolds of the contact pair
-        uint nbToTalContactPoints;
+        uint8 nbToTalContactPoints;
 
         /// True if the colliders of the pair were already colliding in the previous frame
         bool collidingInPreviousFrame;
@@ -90,7 +90,7 @@ struct ContactPair {
 
         /// Constructor
         ContactPair(uint64 pairId, Entity body1Entity, Entity body2Entity, Entity collider1Entity,
-                    Entity collider2Entity, uint contactPairIndex, bool collidingInPreviousFrame, bool isTrigger, MemoryAllocator& allocator)
+                    Entity collider2Entity, uint8 contactPairIndex, bool collidingInPreviousFrame, bool isTrigger, MemoryAllocator& allocator)
             : pairId(pairId), potentialContactManifoldsIndices(allocator), body1Entity(body1Entity), body2Entity(body2Entity),
               collider1Entity(collider1Entity), collider2Entity(collider2Entity),
               isAlreadyInIsland(false), contactPairIndex(contactPairIndex), contactManifoldsIndex(0), nbContactManifolds(0),
